+++ RELEASE_NOTES.md
@@ -7,7 +7,7 @@ https://github.com/padogrid
 
 ### Release Date: 07/03/21
 
-- Added the `open_vscode` command that integrates PadoGrid RWE workspaces with Visual Studio Code.
+- Added the `open_vscode` command that integrates PadoGrid workspaces with Visual Studio Code.
 - Docker cluster support is now compliant with Hazelcast 4.x changes.
 
 ----